implement import menu - cleanup - remove floating action menu - move import logic from GalleryFragment - start moving permission forward to base
