+++ app/src/main/java/dev/leonlatsch/photok/ui/components/base/BaseBottomSheetDialogFragment.kt
@@ -16,28 +16,16 @@
 
 package dev.leonlatsch.photok.ui.components.base
 
-import androidx.appcompat.app.AppCompatActivity
-import androidx.appcompat.widget.Toolbar
-import androidx.fragment.app.Fragment
+import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import pub.devrel.easypermissions.EasyPermissions
 
 /**
- * Base for all fragments.
+ * Base for all bottom sheet dialogs.
  *
  * @since 2.0.0
  * @author Leon Latsch
  */
-abstract class BaseFragment : Fragment() {
-
-    /**
-     * Set the action toolbar for the activity.
-     */
-    internal fun setToolbar(toolbar: Toolbar, showTitle: Boolean = false) {
-        val activity = (requireActivity() as AppCompatActivity)
-        activity.setSupportActionBar(toolbar)
-
-        activity.supportActionBar?.setDisplayShowTitleEnabled(showTitle)
-    }
+abstract class BaseBottomSheetDialogFragment : BottomSheetDialogFragment() {
 
     override fun onRequestPermissionsResult(
         requestCode: Int,
