+++ app/src/main/res/layout/dialog_import_menu.xml
@@ -26,15 +26,18 @@
     <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:orientation="vertical">
+        android:background="@drawable/bg_bottom_sheet"
+        android:orientation="vertical"
+        android:padding="@dimen/medium_gap">
 
         <LinearLayout
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:orientation="horizontal"
-            android:padding="@dimen/medium_gap">
+            android:paddingBottom="@dimen/big_gap">
 
             <LinearLayout
+                android:onClick="@{() -> context.startSelectPhotos()}"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="@dimen/medium_gap"
@@ -43,8 +46,8 @@
                 android:orientation="vertical">
 
                 <ImageView
-                    android:layout_width="60dp"
-                    android:layout_height="60dp"
+                    android:layout_width="45dp"
+                    android:layout_height="45dp"
                     android:src="@drawable/ic_photo_colored" />
 
                 <TextView
@@ -62,8 +65,8 @@
                 android:orientation="vertical">
 
                 <ImageView
-                    android:layout_width="60dp"
-                    android:layout_height="60dp"
+                    android:layout_width="45dp"
+                    android:layout_height="45dp"
                     android:src="@drawable/ic_clapperboard" />
 
                 <TextView
@@ -73,6 +76,7 @@
             </LinearLayout>
 
             <LinearLayout
+                android:onClick="@{() -> context.startSelectBackup()}"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
                 android:layout_marginStart="@dimen/medium_gap"
@@ -81,8 +85,8 @@
                 android:orientation="vertical">
 
                 <ImageView
-                    android:layout_width="60dp"
-                    android:layout_height="60dp"
+                    android:layout_width="45dp"
+                    android:layout_height="45dp"
                     android:src="@drawable/ic_backup" />
 
                 <TextView
@@ -90,9 +94,12 @@
                     android:layout_height="wrap_content"
                     android:text="Backup" />
             </LinearLayout>
-
-
         </LinearLayout>
 
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginStart="16dp"
+            android:text="Select what you want to import" />
     </LinearLayout>
 </layout>
