+++ app/src/main/res/layout/fragment_gallery.xml
@@ -125,44 +125,19 @@
 
             </LinearLayout>
 
-            <com.getbase.floatingactionbutton.FloatingActionsMenu
-                android:id="@+id/galleryActionMenu"
+            <com.google.android.material.floatingactionbutton.FloatingActionButton
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                app:fab_addButtonColorNormal="@color/colorPrimary"
-                app:fab_addButtonColorPressed="@color/colorPrimaryDark"
-                app:fab_labelStyle="@style/FabLabelStyle"
-                app:fab_labelsPosition="left"
+                android:onClick="@{() -> context.showImportMenu()}"
+                android:src="@drawable/ic_add"
+                android:tint="@android:color/white"
                 app:layout_constraintBottom_toBottomOf="parent"
                 app:layout_constraintEnd_toEndOf="parent"
                 app:layout_constraintHorizontal_bias="0.96"
                 app:layout_constraintStart_toStartOf="parent"
                 app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintVertical_bias="0.98">
+                app:layout_constraintVertical_bias="0.98" />
 
-                <com.getbase.floatingactionbutton.FloatingActionButton
-                    android:id="@+id/galleryActionMenuBackup"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:onClick="@{() -> context.startRestore()}"
-                    app:fab_colorNormal="@color/colorPrimaryDark"
-                    app:fab_colorPressed="@android:color/darker_gray"
-                    app:fab_icon="@drawable/ic_backup_restore"
-                    app:fab_size="mini"
-                    app:fab_title="@string/gallery_add_from_backup" />
-
-                <com.getbase.floatingactionbutton.FloatingActionButton
-                    android:id="@+id/galleryActionMenuGallery"
-                    android:layout_width="wrap_content"
-                    android:layout_height="wrap_content"
-                    android:onClick="@{() -> context.startImport()}"
-                    app:fab_colorNormal="@color/colorPrimaryDark"
-                    app:fab_colorPressed="@android:color/darker_gray"
-                    app:fab_icon="@drawable/ic_image_search"
-                    app:fab_size="mini"
-                    app:fab_title="@string/gallery_add_from_gallery" />
-
-            </com.getbase.floatingactionbutton.FloatingActionsMenu>
         </androidx.constraintlayout.widget.ConstraintLayout>
     </LinearLayout>
 </layout>
